Make closing comment syntax more flexible
Before, a space was added before the closing comment syntax (if any).
Now, that is not the case anymore. This allows for using the 'block' or
'mix' syntaxes, without a trailing space in the middle line.

Note that this change is backwards incompatible: A space needs to be
added to the 'universal' syntax in order to be correct.

This also makes it clear that you can add a trailing newline to the
closing comment syntax if your files have to end with a newline.