Also find sourceMappingURL= comments in the middle
not just at the last line of the code. Strictly speaking,
sourceMappingURL= comments have to appear at the end of the file as the
last line. However, for this module to be more useful, I relaxed the
regex a bit. Ideally, we could provide an option to users whether it
should strictly stay with the specification, or pragmatically look at
the entire code for the comments.